+++ src/main/java/dev/redstudio/alfheim/mixin/ChunkCacheMixin.java
@@ -1,7 +1,7 @@
 package dev.redstudio.alfheim.mixin;
 
-import dev.redstudio.alfheim.api.ILightLevelProvider;
 import dev.redstudio.alfheim.api.ILightInfoProvider;
+import dev.redstudio.alfheim.api.ILightLevelProvider;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.ChunkCache;
@@ -9,10 +9,8 @@ import net.minecraft.world.EnumSkyBlock;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 /**
  * @author Luna Lage (Desoroxxx)
@@ -25,10 +23,17 @@ public abstract class ChunkCacheMixin implements ILightLevelProvider {
 
     @Shadow public abstract IBlockState getBlockState(BlockPos pos);
 
+    /**
+     * Explicitly making this method we overwrite public to prevent access level conflicts at runtime.
+     * Mods may adjust the visibility of this method to public, causing a crash if our overwritten method has different scope.
+     *
+     * @reason Redirect to our lighting engine.
+     * @author Luna Lage (Desoroxxx)
+     */
+    @Overwrite
     @SideOnly(Side.CLIENT)
-    @Inject(method = "getLightForExt", at = @At("HEAD"), cancellable = true)
-    private void getLightForExt(final EnumSkyBlock lightType, final BlockPos blockPos, final CallbackInfoReturnable<Integer> callbackInfoReturnable) {
-        callbackInfoReturnable.setReturnValue(((ILightInfoProvider) getBlockState(blockPos)).alfheim$getLightFor(((ChunkCache) (Object) this), lightType, blockPos));
+    public int getLightForExt(final EnumSkyBlock lightType, final BlockPos blockPos) {
+        return ((ILightInfoProvider) getBlockState(blockPos)).alfheim$getLightFor(((ChunkCache) (Object) this), lightType, blockPos);
     }
 
     @Override