+++ Changelog.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Lighting engine will now schedule updates no matter if that chunk is loaded, which is different from vanilla but potentially fix areas lighting being weird when going far away
+- Made all head-cancels overwrite instead
+- All overwrites now make the scope of the overwritten methods `public` to prevent access level conflicts at runtime
 - Now depends on Red Core 0.5
 - Now depends on MixinBooter 8.6
 
@@ -25,6 +27,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Switched to Gradle Kotlin DSL
 - Switched to Adoptium
 - Moved logo to the root of the resources module
+
+### Removed
+
+- Removed access transformed as overwrites can already do the job on their own
 
 ## Alfheim Version 1.0.2 Changelog - 2023-09-12